requirements.txt
Django==4.2.6
django-environ==0.11.2
psycopg2-binary==2.9.9
pandas==2.1.1
Pillow==10.1.0
dependency-injector==4.41.0
djangorestframework==3.14.0
djangorestframework-dataclasses==1.3.1
